Faber creates invitation
