+++ demo/src/Alice.ts
@@ -1,18 +1,15 @@
 import type {
   ConnectionRecord,
-  ConnectionStateChangedEvent,
   CredentialExchangeRecord,
   ProofRecord,
 } from '@aries-framework/core'
-
-import { ConnectionEventTypes } from '@aries-framework/core'
 
 import { BaseAgent } from './BaseAgent'
 import { greenText, Output, redText } from './OutputClass'
 
 export class Alice extends BaseAgent {
-  public outOfBandId?: string
   public connected: boolean
+  public connectionRecordFaberId?: string
 
   public constructor(port: number, name: string) {
     super(port, name)
@@ -26,74 +23,30 @@ export class Alice extends BaseAgent {
   }
 
   private async getConnectionRecord() {
-    if (!this.outOfBandId) {
+    if (!this.connectionRecordFaberId) {
       throw Error(redText(Output.MissingConnectionRecord))
     }
-
-    const [connection] = await this.agent.connections.findAllByOutOfBandId(this.outOfBandId)
-
-    if (!connection) {
-      throw Error(redText(Output.MissingConnectionRecord))
-    }
-
-    return connection
+    return await this.agent.connections.getById(this.connectionRecordFaberId)
   }
 
-  private async printConnectionInvite() {
-    const outOfBand = await this.agent.oob.createInvitation()
-    this.outOfBandId = outOfBand.id
-
-    console.log(
-      Output.ConnectionLink,
-      outOfBand.outOfBandInvitation.toUrl({ domain: `http://localhost:${this.port}` }),
-      '\n'
-    )
+  private async receiveConnectionRequest(invitationUrl: string) {
+    const { connectionRecord } = await this.agent.oob.receiveInvitationFromUrl(invitationUrl)
+    if (!connectionRecord) {
+      throw new Error(redText(Output.NoConnectionRecordFromOutOfBand))
+    }
+    return connectionRecord
   }
 
-  private async waitForConnection() {
-    if (!this.outOfBandId) {
-      throw new Error(redText(Output.MissingConnectionRecord))
-    }
-
-    console.log('Waiting for Faber to finish connection...')
-
-    const getConnectionRecord = (outOfBandId: string) =>
-      new Promise<ConnectionRecord>((resolve, reject) => {
-        // Timeout of 20 seconds
-        const timeoutId = setTimeout(() => reject(new Error(redText(Output.MissingConnectionRecord))), 20000)
-
-        // Start listener
-        this.agent.events.on<ConnectionStateChangedEvent>(ConnectionEventTypes.ConnectionStateChanged, (e) => {
-          if (e.payload.connectionRecord.outOfBandId !== outOfBandId) return
-
-          clearTimeout(timeoutId)
-          resolve(e.payload.connectionRecord)
-        })
-
-        // Also retrieve the connection record by invitation if the event has already fired
-        void this.agent.connections.findAllByOutOfBandId(outOfBandId).then(([connectionRecord]) => {
-          if (connectionRecord) {
-            clearTimeout(timeoutId)
-            resolve(connectionRecord)
-          }
-        })
-      })
-
-    const connectionRecord = await getConnectionRecord(this.outOfBandId)
-
-    try {
-      await this.agent.connections.returnWhenIsConnected(connectionRecord.id)
-    } catch (e) {
-      console.log(redText(`\nTimeout of 20 seconds reached.. Returning to home screen.\n`))
-      return
-    }
-    console.log(greenText(Output.ConnectionEstablished))
+  private async waitForConnection(connectionRecord: ConnectionRecord) {
+    connectionRecord = await this.agent.connections.returnWhenIsConnected(connectionRecord.id)
     this.connected = true
+    console.log(greenText(Output.ConnectionEstablished))
+    return connectionRecord.id
   }
 
-  public async setupConnection() {
-    await this.printConnectionInvite()
-    await this.waitForConnection()
+  public async acceptConnection(invitation_url: string) {
+    const connectionRecord = await this.receiveConnectionRequest(invitation_url)
+    this.connectionRecordFaberId = await this.waitForConnection(connectionRecord)
   }
 
   public async acceptCredentialOffer(credentialRecord: CredentialExchangeRecord) {
